requirements.txt
streamlit>=1.30.0
google-generativeai>=0.8.0
requests>=2.31.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
plotly>=5.18.0
python-dotenv>=1.0.0
lxml>=5.0.0
